LayoutTests/fast/js/kde/const-expected.txt

KDE JS Test

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS c is 11
PASS c is 11
PASS v is 1
PASS successfullyParsed is true

TEST COMPLETE
